probe/jlink: cleanup, remove unused.
